style: rename some files
